device_driver_generation

Module mir

The MIR takes its data from HIR and makes sure all required data is there and all optional data is filled in with defaults.

Modules

kdl_transform

Transforms the MIR back into KDL that should be identical (in information) to the input

lir_transform

passes

Structs

Block

BlockOverride

Buffer

Cfg

Command

CommandOverride

Device

Enum

EnumVariant

Field

GlobalConfig

RefObject

Register

RegisterOverride

Repeat

UniqueId

Enums

Access

BaseType

BitOrder

ByteOrder

EnumGenerationStyle

EnumValue

FieldConversion

Integer

Object

ObjectOverride

ResetValue

Traits

Unique